+++ packages/paraglide/paraglide-js-adapter-sveltekit/src/runtime/config/validatePathTranslations.ts
@@ -1,4 +1,5 @@
-import { parsePathDefinition } from "../path-translations/matching/parse.js"
+import type { ParamMatcher, RouteParam } from "@sveltejs/kit"
+import { parse_route_id } from "../path-translations/matching/routing.js"
 import type { PathTranslations } from "./pathTranslations.js"
 
 export type PathTranslationIssue = {
@@ -12,10 +13,16 @@ export type PathTranslationIssue = {
  */
 export function validatePathTranslations<T extends string>(
 	pathTranslations: PathTranslations<T>,
-	availableLanguageTags: readonly T[]
+	availableLanguageTags: readonly T[],
+	matchers: Record<string, ParamMatcher>
 ): PathTranslationIssue[] {
 	const issues: PathTranslationIssue[] = []
+
+	/** The languages that are available */
 	const expectedLanguages = new Set(availableLanguageTags)
+
+	/** The names of the matchers that are available */
+	const availableMatchers = new Set(Object.keys(matchers))
 
 	for (const path in pathTranslations) {
 		if (!isValidPath(path)) {
@@ -26,7 +33,18 @@ export function validatePathTranslations<T extends string>(
 			continue
 		}
 
-		const expectedParams = getParams(path)
+		const { params: expectedParams } = parse_route_id(path)
+
+		const expectedMatchers = expectedParams.map((param) => param.matcher).filter(Boolean)
+
+		for (const matcher of expectedMatchers) {
+			if (!availableMatchers.has(matcher)) {
+				issues.push({
+					path,
+					message: `Matcher ${matcher} is used but not available. Did you forget to pass it to createI18n?`,
+				})
+			}
+		}
 
 		//@ts-ignore
 		const translations = pathTranslations[path]
@@ -40,7 +58,21 @@ export function validatePathTranslations<T extends string>(
 				})
 			}
 
-			if (!setsEqual(getParams(translatedPath), expectedParams)) {
+			const { params: actualParams } = parse_route_id(translatedPath)
+
+			let paramsDontMatch = false
+
+			for (const param of expectedParams) {
+				if (!actualParams.some((actualParam) => paramsAreEqual(param, actualParam))) {
+					paramsDontMatch = true
+				}
+			}
+
+			if (expectedParams.length !== actualParams.length) {
+				paramsDontMatch = true
+			}
+
+			if (paramsDontMatch) {
 				issues.push({
 					path,
 					message: `The translation for language ${lang} must have the same parameters as the canonical path.`,
@@ -67,30 +99,21 @@ export function validatePathTranslations<T extends string>(
 	return issues
 }
 
+function paramsAreEqual(param1: RouteParam, param2: RouteParam): boolean {
+	return (
+		param1.chained == param2.chained &&
+		param1.matcher == param2.matcher &&
+		param1.name == param2.name &&
+		param1.optional == param2.optional &&
+		param1.rest == param2.rest
+	)
+}
+
 function isValidPath(maybePath: string): maybePath is `/${string}` {
 	return maybePath.startsWith("/")
 }
 
 function isSubset<T>(a: Set<T>, b: Set<T>): boolean {
-	for (const value of a) {
-		if (!b.has(value)) return false
-	}
-	return true
-}
-
-function getParams(path: string): Set<string> {
-	const semgents = parsePathDefinition(path)
-	const params = new Set<string>()
-	for (const segment of semgents) {
-		if (segment.type === "param") {
-			params.add(segment.name)
-		}
-	}
-	return params
-}
-
-function setsEqual(a: Set<any>, b: Set<any>): boolean {
-	if (a.size !== b.size) return false
 	for (const value of a) {
 		if (!b.has(value)) return false
 	}