+++ packages/paraglide/paraglide-js-adapter-sveltekit/src/runtime/path-translations/getCanonicalPath.ts
@@ -1,6 +1,7 @@
-import { matches } from "./matching/match.js"
-import { resolvePath } from "./matching/resolvePath.js"
+import { bestMatch } from "./matching/match.js"
+import { resolve_route } from "./matching/routing.js"
 import type { PathTranslations } from "../config/pathTranslations.js"
+import type { ParamMatcher } from "@sveltejs/kit"
 
 /**
  * Resolves the canonical path from a translated path
@@ -9,7 +10,8 @@ import type { PathTranslations } from "../config/pathTranslations.js"
 export function getCanonicalPath(
 	translatedPath: string,
 	lang: string,
-	translations: PathTranslations
+	translations: PathTranslations,
+	matchers: Record<string, ParamMatcher>
 ): string {
 	for (const [canonicalPathDefinition, translationsForPath] of Object.entries(translations)) {
 		if (!(lang in translationsForPath)) continue
@@ -17,10 +19,10 @@ export function getCanonicalPath(
 		const translatedPathDefinition = translationsForPath[lang]
 		if (!translatedPathDefinition) continue
 
-		const match = matches(translatedPath, [translatedPathDefinition])
+		const match = bestMatch(translatedPath, [translatedPathDefinition], matchers)
 		if (!match) continue
 
-		return resolvePath(canonicalPathDefinition, match.params)
+		return resolve_route(canonicalPathDefinition, match.params)
 	}
 
 	return translatedPath