+++ packages/paraglide/paraglide-js-adapter-sveltekit/src/test.svelte
@@ -1,2 +1,0 @@
-<!-- svelte-ignore a11y-missing-attribute --><a>ASD</a>
-<!-- svelte-ignore a11y-missing-attribute --><a>ASD</a>