+++ packages/settings-component/src/stories/inlang-settings.ts
@@ -8,6 +8,8 @@ import {
 	type InstalledPlugin,
 	type InstalledMessageLintRule,
 } from "@inlang/sdk"
+import checkOptional from "./../helper/checkOptional.js"
+import overridePrimitiveColors from "./../helper/overridePrimitiveColors.js"
 
 import "./input-fields/general-input.js"
 
@@ -16,11 +18,6 @@ import SLOption from "@shoelace-style/shoelace/dist/components/option/option.com
 import SlInput from "@shoelace-style/shoelace/dist/components/input/input.component.js"
 import SlButton from "@shoelace-style/shoelace/dist/components/button/button.component.js"
 import SlCheckbox from "@shoelace-style/shoelace/dist/components/checkbox/checkbox.component.js"
-import SlIconButton from "@shoelace-style/shoelace/dist/components/icon-button/icon-button.component.js"
-import SlIcon from "@shoelace-style/shoelace/dist/components/icon/icon.component.js"
-
-import { setBasePath } from "@shoelace-style/shoelace/dist/utilities/base-path.js"
-setBasePath("./../../../../node_modules/@shoelace-style/shoelace/dist")
 
 // in case an app defines it's own set of shoelace components, prevent double registering
 if (!customElements.get("sl-select")) customElements.define("sl-select", SlSelect)
@@ -28,16 +25,15 @@ if (!customElements.get("sl-option")) customElements.define("sl-option", SLOptio
 if (!customElements.get("sl-input")) customElements.define("sl-input", SlInput)
 if (!customElements.get("sl-button")) customElements.define("sl-button", SlButton)
 if (!customElements.get("sl-checkbox")) customElements.define("sl-checkbox", SlCheckbox)
-if (!customElements.get("sl-icon-button")) customElements.define("sl-icon-button", SlIconButton)
-if (!customElements.get("sl-icon")) customElements.define("sl-icon", SlIcon)
 
 @customElement("inlang-settings")
 export default class InlangSettings extends LitElement {
 	static override styles = [
 		baseStyling,
 		css`
-			h3 {
+			h2 {
 				margin: 0;
+				padding-top: 1rem;
 			}
 			.container {
 				position: relative;
@@ -48,7 +44,7 @@ export default class InlangSettings extends LitElement {
 			.module-container {
 				display: flex;
 				flex-direction: column;
-				gap: 16px;
+				gap: 40px;
 			}
 			.hover-bar-container {
 				width: 100%;
@@ -59,7 +55,8 @@ export default class InlangSettings extends LitElement {
 			.hover-bar {
 				box-sizing: border-box;
 				width: 100%;
-				max-width: 600px;
+				max-width: 500px;
+				height: 48px;
 				margin: 0 auto;
 				display: flex;
 				flex-wrap: wrap;
@@ -67,15 +64,21 @@ export default class InlangSettings extends LitElement {
 				align-items: center;
 				gap: 8px;
 				background-color: var(--sl-panel-background-color);
-				padding: 0.8rem;
 				padding-left: 1rem;
+				padding-right: 0.8rem;
 				border-radius: 0.5rem;
 				border: 1px solid var(--sl-panel-border-color);
 				filter: drop-shadow(0 4px 3px rgb(0 0 0 / 0.07)) drop-shadow(0 2px 2px rgb(0 0 0 / 0.06));
 			}
 			.hover-bar-text {
-				font-weight: 500;
+				font-weight: 600;
+				line-height: 1.5;
+				font-size: 14px;
 				margin: 0;
+			}
+			.test {
+				width: 50px;
+				height: 50px;
 			}
 		`,
 	]
@@ -93,7 +96,7 @@ export default class InlangSettings extends LitElement {
 		[] as ReturnType<InlangProject["installed"]["messageLintRules"]>
 
 	dispatchOnSetSettings(settings: ProjectSettings) {
-		const onSetSettings = new CustomEvent("setSettings", {
+		const onSetSettings = new CustomEvent("set-settings", {
 			detail: {
 				argument: settings,
 			},
@@ -113,6 +116,9 @@ export default class InlangSettings extends LitElement {
 		if (this.settings) {
 			this._newSettings = JSON.parse(JSON.stringify(this.settings))
 		}
+
+		//override primitive colors to match the design system
+		overridePrimitiveColors()
 	}
 
 	handleInlangProjectChange = (
@@ -153,6 +159,7 @@ export default class InlangSettings extends LitElement {
 	_saveChanges = () => {
 		if (this._newSettings) {
 			this.dispatchOnSetSettings(this._newSettings)
+			this.settings = JSON.parse(JSON.stringify(this._newSettings))
 		}
 		this._unsavedChanges = false
 	}
@@ -200,36 +207,44 @@ export default class InlangSettings extends LitElement {
 	}
 
 	override render() {
-		return html` <div class="container">
+		return html` <div class="container" part="base">
 			${Object.entries(this._settingProperties).map(([key, value]) => {
 				return value.schema?.properties && this._newSettings
-					? html`<div class="module-container">
+					? html`<div class="module-container" part="module">
 							${value.meta &&
 							(value.meta?.displayName as { en: string }).en &&
-							html`<h3>${value.meta && (value.meta?.displayName as { en: string }).en}</h3>`}
+							html`<h2 part="module-title">
+								${value.meta && (value.meta?.displayName as { en: string }).en}
+							</h2>`}
 							${Object.entries(value.schema.properties).map(([property, schema]) => {
 								if (property === "$schema" || property === "modules" || property === "experimental")
 									return undefined
 								return key === "internal"
 									? html`
 											<general-input
+												exportparts="property, property-title, property-paragraph, option, option-wrapper, button"
 												.property=${property}
 												.modules=${this.installedMessageLintRules || []}
-												.value=${this._newSettings?.[property as keyof typeof this._newSettings]}
+												.value=${structuredClone(
+													this._newSettings?.[property as keyof typeof this._newSettings]
+												)}
 												.schema=${schema}
 												.handleInlangProjectChange=${this.handleInlangProjectChange}
+												.required=${checkOptional(value.schema, property)}
 											></general-input>
 									  `
 									: html`
 											<general-input
+												exportparts="property, property-title, property-paragraph, option, option-wrapper, button"
 												.property=${property}
 												.value=${
 													// @ts-ignore
-													this._newSettings?.[key]?.[property]
+													structuredClone(this._newSettings?.[key]?.[property])
 												}
 												.schema=${schema}
 												.moduleId=${key}
 												.handleInlangProjectChange=${this.handleInlangProjectChange}
+												.required=${checkOptional(value.schema, property)}
 											></general-input>
 									  `
 							})}
@@ -238,11 +253,12 @@ export default class InlangSettings extends LitElement {
 			})}
 			${this._unsavedChanges
 				? html`<div class="hover-bar-container">
-						<div class="hover-bar">
+						<div class="hover-bar" part="float">
 							<p class="hover-bar-text">Attention, you have unsaved changes.</p>
 							<div>
 								<sl-button
-									size="medium"
+									exportparts="base:button"
+									size="small"
 									@click=${() => {
 										this._revertChanges()
 									}}
@@ -251,7 +267,7 @@ export default class InlangSettings extends LitElement {
 									Cancel
 								</sl-button>
 								<sl-button
-									size="medium"
+									size="small"
 									@click=${() => {
 										this._saveChanges()
 									}}
@@ -273,3 +289,4 @@ declare global {
 		"inlang-settings": InlangSettings
 	}
 }
+