+++ packages/settings-component/src/stories/input-fields/general-input.ts
@@ -1,6 +1,6 @@
 import { html, LitElement } from "lit"
 import { customElement, property } from "lit/decorators.js"
-import { baseStyling } from "../../styling/base.js"
+//import { baseStyling } from "../../styling/base.js"
 import "./string/string-input.js"
 import "./array/array-input.js"
 import "./object/object-input.js"
@@ -8,7 +8,7 @@ import "./union/path-pattern-input.js"
 import type { InstalledMessageLintRule, InstalledPlugin } from "@inlang/sdk"
 @customElement("general-input")
 export class GeneralInput extends LitElement {
-	static override styles = baseStyling
+	//static override styles = baseStyling
 
 	@property()
 	property: string = ""
@@ -26,6 +26,9 @@ export class GeneralInput extends LitElement {
 	schema: any = {}
 
 	@property()
+	required: any = {}
+
+	@property()
 	handleInlangProjectChange: (value: string, key: string, moduleId?: string) => void = () => {}
 
 	override render() {
@@ -33,63 +36,75 @@ export class GeneralInput extends LitElement {
 			if (this.schema.type === "string") {
 				return html` <div>
 					<string-input
+						exportparts="property, property-title, property-paragraph"
 						.property=${this.property}
 						.moduleId=${this.moduleId}
 						.value=${this.value}
 						.schema=${this.schema}
 						.handleInlangProjectChange=${this.handleInlangProjectChange}
+						.required=${this.required}
 					></string-input>
 				</div>`
 			} else if (this.schema.type === "array") {
 				return html` <div>
 					<array-input
+						exportparts="property, property-title, property-paragraph, button"
 						.property=${this.property}
 						.moduleId=${this.moduleId}
 						.value=${this.value}
 						.schema=${this.schema}
 						.handleInlangProjectChange=${this.handleInlangProjectChange}
+						.required=${this.required}
 					></array-input>
 				</div>`
 			} else if (this.schema.type === "object") {
 				return html` <div>
 					<object-input
+						exportparts="property, property-title, property-paragraph, option, option-wrapper, button"
 						.property=${this.property}
 						.moduleId=${this.moduleId}
 						.modules=${this.modules}
 						.value=${this.value}
 						.schema=${this.schema}
 						.handleInlangProjectChange=${this.handleInlangProjectChange}
+						.required=${this.required}
 					></object-input>
 				</div>`
 			} else {
 				return html` <div>
 					<string-input
+						exportparts="property, property-title, property-paragraph"
 						.property=${this.property}
 						.moduleId=${this.moduleId}
 						.value=${this.value}
 						.schema=${this.schema}
 						.handleInlangProjectChange=${this.handleInlangProjectChange}
+						.required=${this.required}
 					></string-input>
 				</div>`
 			}
 		} else if (this.property === "pathPattern" || this.property === "sourceLanguageFilePath") {
 			return html` <div>
 				<path-pattern-input
+					exportparts="property, property-title, property-paragraph, button"
 					.property=${this.property}
 					.moduleId=${this.moduleId}
 					.value=${this.value}
 					.schema=${this.schema}
 					.handleInlangProjectChange=${this.handleInlangProjectChange}
+					.required=${this.required}
 				></path-pattern-input>
 			</div>`
 		} else {
 			return html` <div>
 				<string-input
+					exportparts="property, property-title, property-paragraph"
 					.property=${this.property}
 					.moduleId=${this.moduleId}
 					.value=${this.value}
 					.schema=${this.schema}
 					.handleInlangProjectChange=${this.handleInlangProjectChange}
+					.required=${this.required}
 				></string-input>
 			</div>`
 		}