+++ packages/paraglide/paraglide-js-adapter-next/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @inlang/paraglide-js-adapter-next
 
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies [2428451]
+  - @inlang/paraglide-js@1.5.0
+
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies [d47b2aa]
+- Updated dependencies [192fdec]
+- Updated dependencies [0b7c82e]
+  - @inlang/paraglide-js@1.4.0
+
 ## 1.1.0
 
 ### Minor Changes