+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @inlang/sdk
 
+## 0.31.0
+
+### Minor Changes
+
+- c068dd2: Handle failure to create new project id
+
+## 0.30.0
+
+### Minor Changes
+
+- 9b26a31: Fix saveMessages for non-root projects
+
+## 0.29.0
+
+### Minor Changes
+
+- 62dfa26: SDK Breaking change: LintReports get() and getAll() are now async, and non-reactive.
+  Reduces (does not eliminate) excessive sdk resource consumption.
+
+## 0.28.3
+
+### Patch Changes
+
+- Updated dependencies [8a8aec7]
+  - @lix-js/client@1.2.0
+
 ## 0.28.2
 
 ### Patch Changes