+++ packages/settings-component/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @inlang/settings-component
 
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [c068dd2]
+  - @inlang/sdk@0.31.0
+
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies [9b26a31]
+  - @inlang/sdk@0.30.0
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [62dfa26]
+  - @inlang/sdk@0.29.0
+
+## 0.0.12
+
+### Patch Changes
+
+- @inlang/sdk@0.28.3
+
 ## 0.0.11
 
 ### Patch Changes