+++ packages/website/CHANGELOG.md
@@ -4,6 +4,32 @@
 
 ### Patch Changes
 
+- @lix-js/server@1.1.3
+
+## null
+
+### Patch Changes
+
+- @lix-js/server@1.1.2
+
+## null
+
+### Patch Changes
+
+- @lix-js/server@1.1.1
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [8a8aec7]
+  - @lix-js/client@1.2.0
+  - @lix-js/server@1.1.0
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [d1b361e]
   - @lix-js/client@1.1.0
   - @lix-js/server@1.0.1